feat: added base routes on product
