--- backend/src/models/product/index.ts
@@ -1,2 +1,0 @@
-export { Product } from "./product";
-
