--- backend/src/models/product/type.ts
@@ -1,12 +1,0 @@
-export type TProductImage = {
-    fileName: string;
-    originName: string;
-}
-
-export type TProduct = {
-    title: string;
-    image: TProductImage;
-    category: string;
-    description: string;
-    price: number;
-}
